Replace ModalBottomSheet with Scaffold
