+++ time_counter_flutter_app/pubspec.yaml
@@ -39,7 +39,6 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
   hive_flutter: ^1.1.0
   intl: ^0.17.0
   dynamic_color: ^1.5.4
